refactor(tests): organize unit and integration tests
- Create tests/unit/ for Lambda handler unit tests
- Move Lambda tests with updated import paths
- Keep tests/integration/ for live API tests
- Remove scattered jest.config.js files
