--- backend/lambdas/profile-api/test/events/get-user-comments.json
@@ -1,19 +1,0 @@
-{
-  "httpMethod": "GET",
-  "resource": "/profile/{userId}/comments",
-  "pathParameters": {
-    "userId": "test-user-123"
-  },
-  "queryStringParameters": {
-    "limit": "50"
-  },
-  "requestContext": {
-    "authorizer": {
-      "claims": {
-        "sub": "test-user-123",
-        "email": "test@example.com",
-        "cognito:groups": "ApprovedUsers"
-      }
-    }
-  }
-}
