--- backend/lambdas/profile-api/test/events/update-profile.json
@@ -1,14 +1,0 @@
-{
-  "httpMethod": "PUT",
-  "resource": "/profile",
-  "body": "{\"displayName\":\"John Doe\",\"bio\":\"Family historian and letter enthusiast\",\"familyRelationship\":\"Son of Jane Doe\",\"generation\":\"2nd\",\"familyBranch\":\"Main\",\"isProfilePrivate\":false}",
-  "requestContext": {
-    "authorizer": {
-      "claims": {
-        "sub": "test-user-123",
-        "email": "test@example.com",
-        "cognito:groups": "ApprovedUsers"
-      }
-    }
-  }
-}
