--- backend/lambdas/reactions-api/jest.config.js
@@ -1,11 +1,0 @@
-module.exports = {
-  testEnvironment: 'node',
-  coverageDirectory: 'coverage',
-  collectCoverageFrom: [
-    'index.js',
-    '!test/**'
-  ],
-  testMatch: [
-    '**/test/**/*.test.js'
-  ]
-};
