--- tests/integration/jest.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  testEnvironment: 'node',
-  testTimeout: 30000,
-  testMatch: ['**/*.test.js'],
-  setupFilesAfterEnv: [],
-  verbose: true
-};
